requirements.txt
PyYAML==3.11
docopt==0.6.2
ipython==2.3.1
